Fixing nastiest iphone quirk so far
Turns out if you call focus from another iframe, ios webkit (safari & chrome) does focus on the target but for some reason is frozen. The reason is pretty simple: While visually everything seems focused, the browser is not smart enough to set the focus on the respective frame window. Fuckers.