Implement version check an client updates
This seems a bit premature, but for example Sentry was trashed for days because a user had hiro open in a tab and some extension was throwing errors. So we rather shoud have a way to update outdated clients instead of waiting for weeks until every Chrome crashed/tab was closed/OS rebooted etc.

DONE Add forced version update (with sys.status = online check)