Presumably™ solid client code
Problem: The server always increments the client version.
Example:
- Open two different sessions, write in one.
- Passive session gets first update with sv0 and cv0, sends ack with sv 0 and cv 0, iterates sv to 1 to implictely sync with server.
- Passive session gets second update, this time with CV1 and sv0, client tries to ack with cv0 and sv1, again with sv++ afterwards.
- Passive sessions gets third update, again tries to enforce cv1 but this time also iterated sv to 1. ack is cv0 and sv2, sv3 right after sending that.

Desired behaviour:
Increment server version instead of client version, as diff is done by server and not client. As the client only patches updates, only the server version should be updated on both sides. Incoming changes meaning:
1: sv0 cv0
1ack: sv0 cv0
2: sv 1 cv0
2ack: sv1 cv0

And so on. I think. Or i still didn't get it. Or the contiuum is torn. THE CONTIUUM MIGHT BE TORN!