migrate to SQL backend, implement login-flows
This commit covers the major overhaul of the codebase on the way to
migrating away from gae for our flask-app purposes.
The files in application/ represent a complete independet (from gae
dev env) flask application, which cann be startet with
    $ HIRO_DEBUG=1 python hiro.py
It expects to find a hiro.db in its current dir. (this needs to be
fixed, config values for db-path etc) For now, symlink the hiro.db
sqlite file from hync over here.

Also the login/signup flows have been adapted to cover the definition
and edge-cases written down in the loginflow note.